requirements.txt
# 小红书发布依赖
xhs>=0.2.0
python-dotenv>=1.0.0
requests>=2.28.0